# src/future-be/src/swagger.json
{
	"components": {
...
		"schemas": {
			"AddressRes": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					},
					"default": {
						"type": "boolean"
					}
				},
				"required": [
					"province",
					"district",
					"ward",
					"specificAddress",
					"phone",
					"receiver",
					"_id",
					"default"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateAddressDTO": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					}
				},
				"required": [
					"province",
					"district",
					"ward",
					"specificAddress",
					"phone",
					"receiver"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateAddressDTO_": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateAddressDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_CreateAddressDTO_"
					},
					{
						"properties": {
							"default": {
								"type": "boolean"
							}
						},
						"type": "object"
					}
				]
			},
			"Pick_ICreateAdmin.username-or-name_": {
				"properties": {},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ResAdmin": {
				"$ref": "#/components/schemas/Pick_ICreateAdmin.username-or-name_"
			},
			"Pick_ICreateAdmin.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"AdminResDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_ICreateAdmin.name_"
					},
					{
						"properties": {
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"_id"
						],
						"type": "object"
					}
				]
			},
			"ICreateAdmin": {
				"properties": {
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ICreateAdmin.username-or-password_": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"AdminLogin": {
				"$ref": "#/components/schemas/Pick_ICreateAdmin.username-or-password_"
			},
			"IUpdateAdmin": {
				"properties": {
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"oldPassword": {
						"type": "string"
					}
				},
				"required": [
					"oldPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserAuthenticate": {
				"properties": {
					"token": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ICreateUser.username-or-password_": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Pick_ICreateUser.email_": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IEmailVerify": {
				"$ref": "#/components/schemas/Pick_ICreateUser.email_"
			},
			"Pick_ICreateUser.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserResDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_ICreateUser.name_"
					},
					{
						"properties": {
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"_id"
						],
						"type": "object"
					}
				]
			},
			"ICreateUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"username",
					"password",
					"avatar"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommentResDTO": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"properties": {
							"avatar": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"avatar",
							"name",
							"_id"
						],
						"type": "object"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"content",
					"rate",
					"user",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCommentDTO": {
				"properties": {
					"productId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"productId",
					"content",
					"rate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUserInfo": {
				"properties": {
					"name": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"birthday": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					}
				},
				"required": [
					"name",
					"avatar",
					"email",
					"birthday"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "future-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/addresses": {
			"post": {
				"operationId": "CreateAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/AddressRes"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAddressDTO"
							}
						}
					}
				}
			}
		},
		"/addresses/{id}": {
			"put": {
				"operationId": "UpdateAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/AddressRes"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAddressDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin": {
			"get": {
				"operationId": "GetAdminById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/ResAdmin"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/AdminResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateAdmin"
							}
						}
					}
				}
			}
		},
		"/admin/validate": {
			"get": {
				"operationId": "ValidateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/admin/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminLogin"
							}
						}
					}
				}
			}
		},
		"/admin/setting": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/ResAdmin"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateAdmin"
							}
						}
					}
				}
			}
		},
		"/authenticate/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserAuthenticate"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pick_ICreateUser.username-or-password_"
							}
						}
					}
				}
			}
		},
		"/authenticate/send-code-email": {
			"post": {
				"operationId": "SendVerifyCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEmailVerify"
							}
						}
					}
				}
			}
		},
		"/authenticate/register": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateUser"
							}
						}
					}
				}
			}
		},
...
		"/comments": {
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/CommentResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentDTO"
							}
						}
					}
				}
			}
		},
		"/comments/{productId}/{commentId}": {
			"delete": {
				"operationId": "DeleteComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "commentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateUser"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/IUserInfo"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}